Remove /api path prefix and update dependencies
Since the api gateway is going to be exposed to the internet under
the /api location having the `api` keyword at the microservice
level becomes redundant.

Also I took the freedom to update upload-commons dependency (which
was flagged by github security scans) and the base image to
openjdk 11 (since this jvm is more "container" complaint regarding
resource management).